[DEV-12062] Move React hooks & Intl components from `nextjs` to the `react` package
