+++ packages/nextjs/src/client.ts
@@ -1,2 +1,1 @@
 export * from './adapters/client';
-export * from './hooks';
