+++ packages/nextjs/src/index.ts
@@ -1,3 +1,4 @@
 // Convenience proxy-exports
 export * from '@prezly/theme-kit-core';
 export * from '@prezly/theme-kit-intl';
+export * from '@prezly/theme-kit-react';
