+++ packages/react/src/index.ts
@@ -1,1 +1,2 @@
-export const deleteMe = undefined;
+export * from './hooks';
+export * from './intl';
