+++ packages/react/src/intl/constants.ts
@@ -1,1 +1,2 @@
+export const DEFAULT_LOCALE = 'en';
 export const DEFAULT_TIMEZONE = 'Europe/London';
